requirements.txt
Django==3.2.8
django-rest-framework==0.1.0
pyuploadcare==3.0.0
pydantic==1.9.0
djangorestframework~=3.14.0

httpx~=0.18.2
python-dotenv~=0.19.2